ref: split dynamic config an bounced message tests
